refactor: Member 도메인으로 부터 Auth 도메인 관심사 분리
