+++ src/main/java/com/gomo/app/auth/domain/model/AuthToken.java
@@ -1,0 +1,25 @@
+package com.gomo.app.auth.domain.model;
+
+import com.gomo.app.common.ValueObject;
+
+import lombok.Getter;
+
+@Getter
+@ValueObject
+public class AuthToken {
+
+	private String accessToken;
+	private String refreshToken;
+
+	protected AuthToken() {
+	}
+
+	private AuthToken(String accessToken, String refreshToken) {
+		this.accessToken = accessToken;
+		this.refreshToken = refreshToken;
+	}
+
+	public static AuthToken of(String accessToken, String refreshToken) {
+		return new AuthToken(accessToken, refreshToken);
+	}
+}
