+++ src/main/java/com/gomo/app/auth/domain/repository/AuthTokenRepository.java
@@ -1,0 +1,11 @@
+package com.gomo.app.auth.domain.repository;
+
+import java.util.UUID;
+
+public interface AuthTokenRepository {
+	void setRefreshToken(UUID memberId, String refreshToken);
+
+	String getRefreshToken(UUID memberId);
+
+	void deleteRefreshToken(UUID memberId);
+}
