+++ src/main/java/com/gomo/app/auth/infrastructure/oauth/OAuthProvider.java
@@ -1,0 +1,7 @@
+package com.gomo.app.auth.infrastructure.oauth;
+
+import com.gomo.app.member.domain.model.OAuthUserInfo;
+
+public interface OAuthProvider {
+	OAuthUserInfo authenticate(String code);
+}
