+++ src/main/java/com/gomo/app/auth/presentation/RefreshTokenCookieProvider.java
@@ -1,0 +1,28 @@
+package com.gomo.app.auth.presentation;
+
+import java.time.Duration;
+
+import org.springframework.http.ResponseCookie;
+import org.springframework.stereotype.Service;
+
+@Service
+public class RefreshTokenCookieProvider {
+
+	public ResponseCookie create(String refreshToken, long expiresIn) {
+		return ResponseCookie.from("refreshToken", refreshToken)
+			.httpOnly(true)
+			.secure(true)
+			.path("/")
+			.maxAge(Duration.ofMillis(expiresIn))
+			.build();
+	}
+
+	public ResponseCookie delete() {
+		return ResponseCookie.from("refreshToken", "")
+			.httpOnly(true)
+			.secure(true)
+			.path("/")
+			.maxAge(0)
+			.build();
+	}
+}
