+++ src/main/java/com/gomo/app/auth/presentation/request/CreateEmailAuthCodeRequest.java
@@ -1,0 +1,16 @@
+package com.gomo.app.auth.presentation.request;
+
+import lombok.Getter;
+
+@Getter
+public class CreateEmailAuthCodeRequest {
+	private String email;
+
+	private CreateEmailAuthCodeRequest(String email) {
+		this.email = email;
+	}
+
+	public static CreateEmailAuthCodeRequest of(String email) {
+		return new CreateEmailAuthCodeRequest(email);
+	}
+}
