+++ src/main/java/com/gomo/app/auth/presentation/response/VerifyEmailAuthCodeResponse.java
@@ -1,0 +1,17 @@
+package com.gomo.app.auth.presentation.response;
+
+import lombok.Getter;
+
+@Getter
+public class VerifyEmailAuthCodeResponse {
+
+	private String email;
+
+	private VerifyEmailAuthCodeResponse(String email) {
+		this.email = email;
+	}
+
+	public static VerifyEmailAuthCodeResponse of(String email) {
+		return new VerifyEmailAuthCodeResponse(email);
+	}
+}
