--- src/main/java/com/gomo/app/member/application/RefreshAccessTokenUseCase.java
@@ -1,42 +1,0 @@
-package com.gomo.app.member.application;
-
-import java.util.UUID;
-
-import com.gomo.app.common.ApplicationService;
-import com.gomo.app.common.util.JwtUtil;
-import com.gomo.app.member.exception.MemberAuthenticationFailedException;
-import com.gomo.app.member.exception.code.MemberErrorCode;
-import com.gomo.app.member.infrastructure.JwtSessionRedisService;
-import com.gomo.app.member.presentation.response.LoginMemberResponse;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@ApplicationService
-public class RefreshAccessTokenUseCase {
-
-    private final JwtUtil jwtUtil;
-    private final JwtSessionRedisService jwtSessionRedisService;
-
-    public LoginMemberResponse refresh(String refreshToken){
-        if (refreshToken == null){
-            throw new MemberAuthenticationFailedException(MemberErrorCode.AUTHENTICATION_FAILED);
-        }
-
-        UUID memberId = UUID.fromString(jwtUtil.extractMemberId(refreshToken));
-
-        String storedRefreshToken = jwtSessionRedisService.getRefreshToken(memberId);
-
-        if(!storedRefreshToken.equals(refreshToken)){
-            throw new MemberAuthenticationFailedException(MemberErrorCode.AUTHENTICATION_FAILED);
-        }
-
-        String newAccessToken = jwtUtil.generateAccessToken(memberId);
-        String newRefreshToken = jwtUtil.generateRefreshToken(memberId);
-        long refreshTokenExpiry = jwtUtil.extractExpirationTime(newRefreshToken);
-
-        jwtSessionRedisService.updateRefreshToken(memberId, newRefreshToken);
-
-        return LoginMemberResponse.of(memberId, newAccessToken, newRefreshToken, refreshTokenExpiry);
-    }
-}
