--- src/main/java/com/gomo/app/member/infrastructure/oauth/OAuthProvider.java
@@ -1,7 +1,0 @@
-package com.gomo.app.member.infrastructure.oauth;
-
-import com.gomo.app.member.domain.model.OAuthUserInfo;
-
-public interface OAuthProvider {
-    OAuthUserInfo authenticate(String code);
-}
