--- src/main/java/com/gomo/app/member/presentation/request/VerifyEmailAuthCodeRequest.java
@@ -1,19 +1,0 @@
-package com.gomo.app.member.presentation.request;
-
-import lombok.Getter;
-
-@Getter
-public class VerifyEmailAuthCodeRequest {
-
-    private String email;
-    private String code;
-
-	private VerifyEmailAuthCodeRequest(String email, String code) {
-		this.email = email;
-        this.code = code;
-	}
-
-	public static VerifyEmailAuthCodeRequest of(String email, String code) {
-		return new VerifyEmailAuthCodeRequest(email, code);
-	}
-}
